+++ NEWS.md
@@ -102,6 +102,7 @@ New library features
   data-races. Or use the callback form of `open` to have all that handled
   automatically.
 * `@timed` now additionally returns the elapsed compilation and recompilation time ([#52889])
+* New method `map(f)` is equivalent to `x->map(f,x)` (NOT POSSIBLE UNTIL Julia 2.0 due to [#35293])
 
 Standard library changes
 ------------------------